Update volume slider with additional styling to match player. #11
